+++ Util/src/test/java/org/gusdb/fgputil/db/platform/DBPlatformTest.java
@@ -14,9 +14,9 @@ public class DBPlatformTest {
     @Override public String getConnectionUrl() { return "url://mydb"; }
     @Override public SupportedPlatform getPlatformEnum() { return SupportedPlatform.ORACLE; }
     @Override public String getDriverInitClass() { return null; }
-    @Override public short getMaxActive() { return 20; }
-    @Override public short getMaxIdle() { return 20; }
-    @Override public short getMinIdle() { return 5; }
+    @Override public int getMaxActive() { return 20; }
+    @Override public int getMaxIdle() { return 20; }
+    @Override public int getMinIdle() { return 5; }
     @Override public long getMaxWait() { return 2000; }
     @Override public int getDefaultFetchSize() { return 0; }
     @Override public boolean isShowConnections() { return true; }
